+++ common/util.go
@@ -272,7 +272,7 @@ func PriceToFixed(price *big.Rat) (int64, error) {
 	return ratToFixed(price, priceScalingFactor)
 }
 
-// FixedToPrice converts an fixed point number with 3 decimal places represented as in int64 into a big.Rat
+// FixedToPrice converts a fixed point number with 3 decimal places represented as in int64 into a big.Rat
 func FixedToPrice(price int64) *big.Rat {
 	return big.NewRat(price, priceScalingFactor)
 }