+++ core/livepeernode.go
@@ -155,10 +155,17 @@ type LivepeerNode struct {
 
 	MediaMTXApiPassword        string
 	LiveAITrickleHostForRunner string
+	LiveAIAuthApiKey           string
+	LivePaymentInterval        time.Duration
+
+	// Gateway
+	GatewayHost string
 }
 
 type LivePipeline struct {
-	ControlPub *trickle.TricklePublisher
+	ControlPub  *trickle.TricklePublisher
+	StopControl func()
+	EventsPub   *trickle.TricklePublisher
 }
 
 // NewLivepeerNode creates a new Livepeer Node. Eth can be nil.