+++ server/ai_process.go
@@ -88,11 +88,30 @@ type aiRequestParams struct {
 	liveParams liveRequestParams
 }
 
+func (a aiRequestParams) inputStreamExists() bool {
+	if a.node == nil {
+		return false
+	}
+	_, ok := a.node.LivePipelines[a.liveParams.stream]
+	return ok
+}
+
 // For live video pipelines
 type liveRequestParams struct {
 	segmentReader *media.SwitchableSegmentReader
 	outputRTMPURL string
 	stream        string
+	requestID     string
+	streamID      string
+	pipelineID    string
+
+	paymentProcessInterval time.Duration
+
+	// Stops the pipeline with an error. Also kicks the input
+	stopPipeline func(error)
+
+	// Report an error event
+	sendErrorEvent func(error)
 }
 
 // CalculateTextToImageLatencyScore computes the time taken per pixel for an text-to-image request.
@@ -1004,7 +1023,14 @@ func submitAudioToText(ctx context.Context, params aiRequestParams, sess *AISess
 	return &res, nil
 }
 
+const initPixelsToPay = 30 * 30 * 3200 * 1800 // 30 seconds, 30fps, 1800p
+
 func submitLiveVideoToVideo(ctx context.Context, params aiRequestParams, sess *AISession, req worker.GenLiveVideoToVideoJSONRequestBody) (any, error) {
+	// Live Video should not reuse the existing session balance, because it could lead to not sending the init
+	// payment, which in turns may cause "Insufficient Balance" on the Orchestrator's side.
+	// It works differently than other AI Jobs, because Live Video is accounted by mid on the Orchestrator's side.
+	clearSessionBalance(sess.BroadcastSession, core.RandomManifestID())
+
 	client, err := worker.NewClientWithResponses(sess.Transcoder(), worker.WithHTTPClient(httpClient))
 	if err != nil {
 		if monitor.Enabled {
@@ -1012,32 +1038,57 @@ func submitLiveVideoToVideo(ctx context.Context, params aiRequestParams, sess *A
 		}
 		return nil, err
 	}
+	setHeaders, balUpdate, err := prepareAIPayment(ctx, sess, initPixelsToPay)
+	defer completeBalanceUpdate(sess.BroadcastSession, balUpdate)
 
 	// Send request to orchestrator
-	resp, err := client.GenLiveVideoToVideoWithResponse(ctx, req)
+	resp, err := client.GenLiveVideoToVideoWithResponse(ctx, req, setHeaders)
 	if err != nil {
 		return nil, err
 	}
 
-	if resp.JSON200 != nil {
-		host := sess.Transcoder()
-		pub, err := common.AppendHostname(resp.JSON200.PublishUrl, host)
-		if err != nil {
-			return nil, fmt.Errorf("invalid publish URL: %w", err)
-		}
-		sub, err := common.AppendHostname(resp.JSON200.SubscribeUrl, host)
-		if err != nil {
-			return nil, fmt.Errorf("invalid subscribe URL: %w", err)
-		}
-		control, err := common.AppendHostname(resp.JSON200.ControlUrl, host)
-		if err != nil {
-			return nil, fmt.Errorf("invalid control URL: %w", err)
-		}
-		startTricklePublish(pub, params)
-		startTrickleSubscribe(sub, params)
-		startControlPublish(control, params)
+	if resp.JSON200 == nil {
+		// TODO: Replace trim newline with better error spec from O
+		return nil, errors.New(strings.TrimSuffix(string(resp.Body), "\n"))
 	}
+
+	if resp.JSON200.ControlUrl == nil {
+		return nil, errors.New("control URL is missing")
+	}
+
+	host := sess.Transcoder()
+	pub, err := common.AppendHostname(resp.JSON200.PublishUrl, host)
+	if err != nil {
+		return nil, fmt.Errorf("invalid publish URL: %w", err)
+	}
+	sub, err := common.AppendHostname(resp.JSON200.SubscribeUrl, host)
+	if err != nil {
+		return nil, fmt.Errorf("invalid subscribe URL: %w", err)
+	}
+	control, err := common.AppendHostname(*resp.JSON200.ControlUrl, host)
+	if err != nil {
+		return nil, fmt.Errorf("invalid control URL: %w", err)
+	}
+	events, err := common.AppendHostname(*resp.JSON200.EventsUrl, host)
+	if err != nil {
+		return nil, fmt.Errorf("invalid events URL: %w", err)
+	}
+	clog.V(common.VERBOSE).Infof(ctx, "pub %s sub %s control %s events %s", pub, sub, control, events)
+
+	startControlPublish(control, params)
+	startTricklePublish(ctx, pub, params, sess)
+	startTrickleSubscribe(ctx, sub, params)
+	startEventsSubscribe(ctx, events, params, sess)
 	return resp, nil
+}
+
+// extractMid extracts the mid (manifest ID) from the publish URL
+// e.g. public URL passed from orchestrator: /live/manifest/123456, then mid is 123456
+// we can consider improving it and passing mid directly in the JSON response from Orchestrator,
+// but currently it would require changing the OpenAPI schema in livepeer/ai-worker repo
+func extractMid(path string) string {
+	pubSplit := strings.Split(path, "/")
+	return pubSplit[len(pubSplit)-1]
 }
 
 func CalculateLLMLatencyScore(took time.Duration, tokensUsed int) float64 {
@@ -1393,6 +1444,9 @@ func processAIRequest(ctx context.Context, params aiRequestParams, req interface
 		modelID = defaultLiveVideoToVideoModelID
 		if v.ModelId != nil && *v.ModelId != "" {
 			modelID = *v.ModelId
+		} else {
+			// set default model
+			v.ModelId = &modelID
 		}
 		submitFn = func(ctx context.Context, params aiRequestParams, sess *AISession) (interface{}, error) {
 			return submitLiveVideoToVideo(ctx, params, sess, v)