+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## v0.8.1
+
+- [#3279](https://github.com/livepeer/go-livepeer/pull/3279) Enable automatic worker image pulling.
+- Live Video AI features
+- [#3249](https://github.com/livepeer/go-livepeer/pull/3249) Add Gateway ETH Address to Kafka events
+
+### Features ⚒
+
+#### Broadcaster
+
+### Bug Fixes 🐞
+
+#### Broadcaster
+
 ## v0.8.0
 
 - [#2959](https://github.com/livepeer/go-livepeer/pull/2959) Add Livepeer AI Subnet features